kafka-server/requirements.txt
kafka-python==2.0.2
flask==3.0.0
boto3==1.34.21